add ip_addresses shell script
The ip_addresses shell script contains regular expressions for valid
IPv4 and IPv6 addresses. Those are tested against a collection of
addresses and printed to STDOUT if the tests all pass.

The regular expressions are intended to only match valid IP addresses,
as opposed to similarly structured strings that are not syntactically
and semantically valid representations of unsigned 32bit or 128bit
integer values as used for IP addressing.

As an example, the regular expression for IPv4 addresses shall match
"1.2.3.4", but shall not match "1.256.3.4".