add three "nz" implementations: nz.awk, nz.sed, nz.sh
The "nz" filter suppresses lines that contain decimal numbers, if
all those numbers are zero. This can be useful if a list of error
counters contains mostly zero values, but one is interested in the
few non-zero cases.

The simple "nz" filter has no special cases, so that e.g. an
interface name of "eth0" or "Ethernet0/0" would be filtered out
as well.