split lcunfold2 into lcunfold2-1 and lcunfold2-2
The Internet-Draft "Handling Long Lines in Inclusions in
Internet-Drafts and RFCs" specifies two line wrapping
strategies, because the second can correctly wrap text
that 1) contains a line ending with a backslash ('\'),
and 2) contains a line that has whitespace after the
wrapping position.

Combining both strategies in one unfolding script, as was
done in lcunfold2, results in incorrectly re-combining
lines if strategy 2 was needed because of case 1) above.