add shell script striprfcpagination
This script removes pagination artifacts from an RFC or Internet-Draft
in text format. The result is well suited to be read with a common
pager such as more or less using a normal size terminal.

Reading the original RFC respectively Internet-Draft in text format
using a pager requires firstly an oversized terminal and secondly a
pager that interprets form-feed ('\f') control characters as filling
the current page, which neither more nor less do by default.

Note that "more -p" and "more -c" can be used to read the original RFC
or Internet-Draft text one page at a time, if the terminal is large
enough.