add Python script resize.py for EXOS switches
The Python script `resize.py` can be used on switches using the
network operating system ExtremeXOS (EXOS) to set the logical
terminal size used by the CLI to the size of the used terminal.
This works for remote sessions, e.g., using SSH or Telnet, but
not for the serial console.

The script uses a Python module available on EXOS switches with
Python, and uses that to execute EXOS specific CLI commands on
the switch.