dhms: mawk compatibility part 2
mawk 1.3.4 20200120 evaluates the locale (LC_NUMERIC) settings
and uses the locale specific decimal separator for floating
point in- and output.  With LC_NUMERIC=de_DE.UTF-8 as I use
this means a comma (',') instead of a dot ('.').  This commit
allows input to use either '.' or ',' as decimal separator,
but only the correct character for the combination of Awk and
LC_NUMERIC setting results in functioning floating point input.
Using the wrong decimal separator results in silently ignored
fractional values.

This completes mawk compatibility for 'dhms'.