dhms: reset total to 0 before converting DHMS to s
This is required to correctly convert more than one DHMS value
to seconds, i.e., when dhms is given more than one input line
with a DHMS value.

This is a de-facto revert of 45dde98176c6d4b2af6e81c40b8e6ad4b64921a8
with an added comment to hint atthe fact that this code line is
required for correctness.