dhms: accept input in generalized HH:MM:SS format
The first portion of the input in HH:MM:SS format, optionally
with more than two digits for hours, optionally only one digit
for minutes or seconds, and optionally fractional seconds, is
converted into "hms" format with units.  This transformed input
is later checked for conformance to a generalized "dhms" format.