ip_addresses: more test cases
On an IPv6-related IETF mailing lists, it was reported[1] that some
implementation supports zone identifiers with the IPv6 loopback address.
GNU/Linux was reported[2] as not supporting it, which I verified myself.
The ip_addresses script does not accept a zone identifier with the IPv6
loopback address.

RFC 4291[3] states:

  It must not be assigned to any physical interface.  It is treated
  as having Link-Local scope, and may be thought of as the Link-Local
  unicast address of a virtual interface (typically called the "loopback
  interface") to an imaginary link that goes nowhere.

While Link-Local scope suggests that a zone identifier might be OK, not
assigning it to any interface make any zone identifier incorrect, because
none can identify an existing interface with this address assigned.
But that is useless rule interpretation.  Reality is that different
systems implement this differently.

I arbitrarily decide the current ip_addresses implementation is OK,
and add scoped IPv6 loopback addresses as non-address test cases.
This at least documents the current behavior of the regular expressions
in the script.

[1] https://mailarchive.ietf.org/arch/msg/ipv6/kweOR4hnmX2UPL-QaGO9vR85BmU/
[2] https://mailarchive.ietf.org/arch/msg/ipv6/G95_CyLTlhLimsvHYJ_dlEmWCQ8/
[3] https://www.rfc-editor.org/rfc/rfc4291#section-2.5.3